+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to **Pipecat** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-05-15
+
+### Changed
+
+- Updated NVIDIA SageMaker STT and TTS generated service configs to use
+  `AWS_REGION` for the SageMaker endpoint region.
+
 ## [1.2.0] - 2026-05-15
 
 ### Added